Show amount of happenings in sidebar and move fetching to page component
