+++ apps/web/src/app/(default)/for-studenter/arrangementer/page.tsx
@@ -1,4 +1,5 @@
 import { Suspense } from "react";
+import { nextMonday, subMinutes } from "date-fns";
 
 import { Container } from "@/components/container";
 import {
@@ -7,10 +8,26 @@ import {
   EventSearchAndOrderBar,
 } from "@/components/event-filter";
 import EventsView from "@/components/events-view";
+import { fetchFilteredHappening } from "@/sanity/happening/requests";
+
+/**
+ * This is a type definition for the query that is sent to Sanity.
+ */
+export type FilteredHappeningQuery = {
+  search?: string;
+  type: "all" | "event" | "bedpres";
+  open: boolean;
+  dateFilter?: Array<DateInterval>;
+};
+
+type DateInterval = {
+  start?: Date;
+  end?: Date;
+};
 
 export type SearchParams = {
   type: string;
-  order: string;
+  order?: string;
   search?: string;
   open?: string;
   past?: string;
@@ -19,11 +36,84 @@ export type SearchParams = {
   later?: string;
 };
 
-export default function Page({ searchParams }: { searchParams?: SearchParams }) {
+// Sanitizes the query params before fetching data
+function generateQuery(params: SearchParams) {
+  const query: FilteredHappeningQuery = {
+    search: params.search ?? undefined,
+    type: (params.type as "event" | "bedpres" | "all") ?? "all",
+    open: params.open === "true" ? true : false,
+    dateFilter: getDateInterval(params),
+  };
+
+  return query;
+}
+
+function getDateInterval(params: SearchParams) {
+  const currentDate = new Date();
+
+  const past = params.past === "true" ? true : false;
+  const hideThisWeek = params.thisWeek === "false" ? true : false;
+  const hideNextWeek = params.nextWeek === "false" ? true : false;
+  const hideLater = params.later === "false" ? true : false;
+
+  if (past) {
+    return [{ end: currentDate }];
+  }
+  if (!hideThisWeek && !hideNextWeek && !hideLater) return [{ start: subMinutes(currentDate, 30) }];
+  if (hideThisWeek && !hideNextWeek && !hideLater) return [{ start: nextMonday(currentDate) }];
+  if (!hideThisWeek && hideNextWeek && !hideLater)
+    return [
+      { start: subMinutes(currentDate, 30), end: nextMonday(currentDate) },
+      { start: nextMonday(nextMonday(currentDate)) },
+    ];
+  if (!hideThisWeek && !hideNextWeek && hideLater)
+    return [{ start: subMinutes(currentDate, 30), end: nextMonday(nextMonday(currentDate)) }];
+  if (!hideThisWeek && hideNextWeek && hideLater)
+    return [{ start: subMinutes(currentDate, 30), end: nextMonday(currentDate) }];
+  if (hideThisWeek && !hideNextWeek && hideLater)
+    return [{ start: nextMonday(currentDate), end: nextMonday(nextMonday(currentDate)) }];
+  if (hideThisWeek && hideNextWeek && !hideLater)
+    return [{ start: nextMonday(nextMonday(currentDate)) }];
+
+  return undefined;
+}
+
+export default async function Page({ searchParams }: { searchParams?: SearchParams }) {
   const params = searchParams ?? {
     type: "all",
-    order: "DESC",
   };
+
+  const query = generateQuery(params);
+  const happenings = await fetchFilteredHappening(query);
+
+  if (params.order === "ASC") happenings.reverse();
+
+  const { numThisWeek, numNextWeek, numLater } = happenings.reduce(
+    (acc: { numThisWeek: number; numNextWeek: number; numLater: number }, happening) => {
+      const { date } = happening;
+
+      if (!date) {
+        return acc;
+      }
+
+      const happeningDate = new Date(date);
+
+      const thisWeek = subMinutes(new Date(), 30);
+      const nextWeek = nextMonday(thisWeek);
+      const later = nextMonday(nextWeek);
+
+      if (happeningDate >= thisWeek && happeningDate < nextWeek) {
+        acc.numThisWeek += 1;
+      } else if (happeningDate >= nextWeek && happeningDate < later) {
+        acc.numNextWeek += 1;
+      } else if (happeningDate >= later) {
+        acc.numLater += 1;
+      }
+
+      return acc;
+    },
+    { numThisWeek: 0, numNextWeek: 0, numLater: 0 },
+  );
 
   return (
     <Container>
@@ -36,11 +126,11 @@ export default function Page({ searchParams }: { searchParams?: SearchParams }) 
         </div>
         <div>
           <div>
-            <EventDateFilterSidebar />
+            <EventDateFilterSidebar numOfEvents={{ numThisWeek, numNextWeek, numLater }} />
           </div>
           <div>
             <Suspense fallback={<p>Laster...</p>}>
-              <EventsView params={params} />
+              <EventsView happenings={happenings} />
             </Suspense>
           </div>
         </div>
